:art: modularize code
