+++ package.json
@@ -5,7 +5,6 @@
     "url": "http://github.com/ricmoo/aes-js/issues",
     "email": "github@ricmoo.com"
   },
-  "type": "module",
   "description": "A pure JavaScript implementation of the AES block cipher and all common modes of operation.",
   "devDependencies": {
     "esm": "^3.2.25",
