+++ tsconfig.json
@@ -1,6 +1,7 @@
 {
-  "include": ["src/index.ts"],
+  "include": ["src"],
   "compilerOptions": {
+    "isolatedModules": true,
     "moduleResolution": "node",
     "module": "es6",
     "strict": true,
